findbugs
git-svn-id: svn://svn.code.sf.net/p/pmii/code/trunk@4651 2dc5e0b7-7910-0410-9abd-fb16afba6634

+++ QedeqBase/src/org/qedeq/base/io/UrlUtility.java
@@ -221,7 +221,7 @@ public final class UrlUtility {
                 throw new FileNotFoundException("\"" + url + "\" was substituted by "
                     + "\"" + connection.getURL() + "\" from server");
             }
-            final int maximum = connection.getContentLength();
+            final double maximum = connection.getContentLength();
             IoUtility.createNecessaryDirectories(f);
             out = new FileOutputStream(f);
             final byte[] buffer = new byte[4096];
